requirements/base.txt
# Base dependencies go here

# Bleeding edge Django
# -----------------------------------------------------------------------------
django==1.9

# Configuration
# -----------------------------------------------------------------------------
django-environ

# Images
# -----------------------------------------------------------------------------
Pillow

# Admin
# -----------------------------------------------------------------------------
django-suit

# Your custom requirements go here
# -----------------------------------------------------------------------------
djangorestframework==3.5.3